newsorbit/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pymongo==4.8.0
firebase-admin==6.5.0
google-generativeai==0.8.2
httpx==0.27.2
APScheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.5.2
